requirements.txt
streamlit>=1.32.0
openai>=1.12.0
msal>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.5
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
